requirements.txt
nginx
Copy
Edit
Django
pandas
numpy
matplotlib
seaborn
faiss-cpu
requests